Restructuring folder organization. #18
