+++ app-paper/src/main/java/com/paper/editor/widget/canvas/IBaseWidget.kt
@@ -18,7 +18,7 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-package com.paper.editor.widget
+package com.paper.editor.widget.canvas
 
 interface IBaseWidget<in T> {
